move codable feed store into its own file
